Update Assignment_1.qmd
Updated author names
Added the interpretations for the question 3
Added R codes for reference

There are 2 things that require revisit:
 b-> should we accept normality (I wrote both answers one rejecting one agreeing, accordingly to the length of the document I can update/lengthify it)
d->, I used the example on the Lecture 2 slide page 17. However, sample proportion was not given in the current question while it was with the other example. I used 0.5 for maximum variability. So phat*(1-phat) was used for calculation. It got the result 2401, and in the book when its unknown says instead of using phat*(1-phat), it is advised to use 0.25 (instead of(p*q) value. Again it gave 2401 as a result. Just wanted to add it as an additional information.